modify: replace printf and GNL folders in libft
