--- libft/libft/Makefile
@@ -1,67 +1,0 @@
-# **************************************************************************** #
-#                                                                              #
-#                                                         :::      ::::::::    #
-#    Makefile                                           :+:      :+:    :+:    #
-#                                                     +:+ +:+         +:+      #
-#    By: chajung <chajung@student.42seoul.kr>       +#+  +:+       +#+         #
-#                                                 +#+#+#+#+#+   +#+            #
-#    Created: 2022/11/09 20:35:22 by chajung           #+#    #+#              #
-#    Updated: 2022/11/21 18:43:54 by chajung          ###   ########.fr        #
-#                                                                              #
-# **************************************************************************** #
-
-.DEFAULT_GOAL := all
-
-NAME		= libft.a
-
-CC			= cc
-CFLAGS		= -Wall -Wextra -Werror
-AR			= ar
-RCS			= rcs
-RMFLAGS		= -r
-
-SRC_DIR		= src
-OBJ_DIR		= obj
-
-SRC 		= ft_memset.c ft_isalnum.c ft_isalpha.c ft_isascii.c ft_isdigit.c \
-			  ft_isdigit.c ft_isprint.c ft_tolower.c ft_toupper.c ft_atoi.c \
-			  ft_strncmp.c ft_strlen.c ft_strlcpy.c ft_strlcat.c ft_strchr.c \
-			  ft_strrchr.c ft_substr.c ft_strnstr.c ft_calloc.c ft_strdup.c \
-			  ft_strjoin.c ft_itoa.c ft_strtrim.c ft_split.c ft_bzero.c ft_memcpy.c \
-			  ft_memmove.c ft_memchr.c ft_memcmp.c ft_strmapi.c ft_striteri.c \
-			  ft_putchar_fd.c ft_putendl_fd.c ft_putnbr_fd.c ft_putstr_fd.c \
-			  ft_lstnew_bonus.c ft_lstadd_front_bonus.c ft_lstsize_bonus.c \
-			  ft_lstlast_bonus.c ft_lstadd_back_bonus.c ft_lstdelone_bonus.c \
-			  ft_lstclear_bonus.c ft_lstiter_bonus.c ft_lstmap_bonus.c
-
-SRCS		= $(addprefix $(SRC_DIR)/, $(SRC))
-
-OBJ 		= $(SRC:.c=.o) 
-OBJS		= $(addprefix $(OBJ_DIR)/, $(OBJ))
-
-all: $(OBJ_DIR) $(NAME)
-
-$(NAME): $(OBJS)
-	@$(AR) $(RCS) $@ $^
-	@echo "libft make done"
-
-$(OBJ_DIR):
-	@mkdir -p $(@)
-
-$(OBJ_DIR)/%.o: $(SRC_DIR)/%.c
-	@$(CC) $(CFLAGS) -c $< -o $@
-
-clean:
-	@$(RM) $(RMFLAGS) $(OBJ_DIR)
-	@echo "libft clean done"
-
-fclean:
-	@$(RM) $(RMFLAGS) $(OBJ_DIR)
-	@$(RM) $(RMFLAGS) $(NAME)
-	@echo "libft fclean done"
-
-re:
-	@$(MAKE) fclean
-	@$(MAKE) all
-
-.PHONY: all clean fclean re
