+++ libft/src/ft_lstiter_bonus.c
@@ -1,31 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strmapi.c                                       :+:      :+:    :+:   */
+/*   ft_lstiter.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: chajung <chajung@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/11/10 14:39:26 by chajung           #+#    #+#             */
-/*   Updated: 2022/11/13 14:08:44 by chajung          ###   ########.fr       */
+/*   Created: 2022/11/12 16:42:15 by chajung           #+#    #+#             */
+/*   Updated: 2022/11/13 14:27:12 by chajung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../inc/libft.h"
+#include "../inc/libft.h"
 
-char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
+void	ft_lstiter(t_list *lst, void (*f)(void *))
 {
-	unsigned int	i;
-	char			*addr;
-
-	addr = (char *)malloc(sizeof(char) * (ft_strlen(s) + 1));
-	if (addr == NULL)
-		return (NULL);
-	i = 0;
-	while (s[i])
+	while (lst)
 	{
-		addr[i] = f(i, s[i]);
-		i++;
+		f(lst->content);
+		lst = lst->next;
 	}
-	addr[i] = '\0';
-	return (addr);
 }
