+++ libft/src/ft_lstlast_bonus.c
@@ -1,21 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_lstadd_back.c                                   :+:      :+:    :+:   */
+/*   ft_lstlast.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: chajung <chajung@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/11/12 16:39:17 by chajung           #+#    #+#             */
-/*   Updated: 2022/11/21 14:53:58 by chajung          ###   ########.fr       */
+/*   Created: 2022/11/12 16:40:25 by chajung           #+#    #+#             */
+/*   Updated: 2022/11/13 14:27:14 by chajung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../inc/libft.h"
+#include "../inc/libft.h"
 
-void	ft_lstadd_back(t_list **lst, t_list *new)
+t_list	*ft_lstlast(t_list *lst)
 {
-	if (*lst == NULL)
-		*lst = new;
-	else
-		ft_lstlast(*lst)->next = new;
+	while (lst && lst->next)
+		lst = lst->next;
+	return (lst);
 }
