+++ libft/src/ft_lstnew_bonus.c
@@ -1,20 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_lstlast.c                                       :+:      :+:    :+:   */
+/*   ft_lstnew.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: chajung <chajung@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/11/12 16:40:25 by chajung           #+#    #+#             */
-/*   Updated: 2022/11/13 14:27:14 by chajung          ###   ########.fr       */
+/*   Created: 2022/11/12 15:26:38 by chajung           #+#    #+#             */
+/*   Updated: 2022/11/13 14:27:16 by chajung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../inc/libft.h"
+#include "../inc/libft.h"
 
-t_list	*ft_lstlast(t_list *lst)
+t_list	*ft_lstnew(void *content)
 {
-	while (lst && lst->next)
-		lst = lst->next;
+	t_list	*lst;
+
+	lst = (t_list *)malloc(sizeof(t_list));
+	if (lst == NULL)
+		return (NULL);
+	lst->content = content;
+	lst->next = NULL;
 	return (lst);
 }
