+++ libft/src/ft_memchr.c
