+++ libft/src/ft_putendl_fd.c
@@ -1,26 +1,19 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_lstsize.c                                       :+:      :+:    :+:   */
+/*   ft_putendl_fd.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: chajung <chajung@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/11/12 17:13:30 by chajung           #+#    #+#             */
-/*   Updated: 2022/11/13 14:27:21 by chajung          ###   ########.fr       */
+/*   Created: 2022/11/10 17:00:18 by chajung           #+#    #+#             */
+/*   Updated: 2022/11/10 17:32:49 by chajung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../inc/libft.h"
+#include "../inc/libft.h"
 
-int	ft_lstsize(t_list *lst)
+void	ft_putendl_fd(char *s, int fd)
 {
-	int		size;
-
-	size = 0;
-	while (lst)
-	{
-		lst = lst->next;
-		size++;
-	}
-	return (size);
+	ft_putstr_fd(s, fd);
+	ft_putchar_fd('\n', fd);
 }
