+++ libft/src/ft_putnbr_fd.c
@@ -1,29 +1,39 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strlcpy.c                                       :+:      :+:    :+:   */
+/*   ft_putnbr_fd.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: chajung <chajung@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/11/09 18:10:57 by chajung           #+#    #+#             */
-/*   Updated: 2022/11/13 14:08:35 by chajung          ###   ########.fr       */
+/*   Created: 2022/11/10 17:04:33 by chajung           #+#    #+#             */
+/*   Updated: 2022/11/13 14:07:01 by chajung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../inc/libft.h"
+#include "../inc/libft.h"
 
-size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
+static void	put_nbr_recursive(int n, int fd)
 {
-	size_t	i;
-
-	if (dstsize == 0)
-		return (ft_strlen(src));
-	i = 0;
-	while (src[i] && i + 1 < dstsize)
+	if (n < 10)
+		ft_putchar_fd(n + '0', fd);
+	else
 	{
-		dst[i] = src[i];
-		i++;
+		put_nbr_recursive(n / 10, fd);
+		put_nbr_recursive(n % 10, fd);
 	}
-	dst[i] = '\0';
-	return (ft_strlen(src));
+}
+
+void	ft_putnbr_fd(int n, int fd)
+{
+	if (n != -2147483648)
+	{
+		if (n < 0)
+		{
+			ft_putchar_fd('-', fd);
+			n *= -1;
+		}
+		put_nbr_recursive(n, fd);
+	}
+	else
+		ft_putstr_fd("-2147483648", fd);
 }
