+++ libft/src/ft_strdup.c
@@ -1,29 +1,28 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strrchr.c                                       :+:      :+:    :+:   */
+/*   ft_strdup.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: chajung <chajung@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/11/09 18:42:01 by chajung           #+#    #+#             */
-/*   Updated: 2022/11/13 14:09:08 by chajung          ###   ########.fr       */
+/*   Created: 2022/11/09 19:56:32 by chajung           #+#    #+#             */
+/*   Updated: 2022/11/13 14:08:13 by chajung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../inc/libft.h"
+#include "../inc/libft.h"
 
-char	*ft_strrchr(const char *s, int c)
+char	*ft_strdup(const char *s1)
 {
-	char	*temp;
+	char	*addr;
+	int		i;
 
-	temp = NULL;
-	if ((char)c == '\0')
-		return ((char *)s + ft_strlen(s));
-	while (*s)
-	{
-		if (*s == (char)c)
-			temp = (char *)s;
-		s++;
-	}
-	return (temp);
+	addr = (char *)malloc(sizeof(char) * (ft_strlen(s1) + 1));
+	if (addr == NULL)
+		return (NULL);
+	i = -1;
+	while (s1[++i])
+		addr[i] = s1[i];
+	addr[i] = '\0';
+	return (addr);
 }
