+++ libft/src/ft_striteri.c
@@ -1,18 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_putstr_fd.c                                     :+:      :+:    :+:   */
+/*   ft_striteri.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: chajung <chajung@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/11/10 16:57:57 by chajung           #+#    #+#             */
-/*   Updated: 2022/11/10 17:00:02 by chajung          ###   ########.fr       */
+/*   Created: 2022/11/10 14:53:26 by chajung           #+#    #+#             */
+/*   Updated: 2022/11/10 17:35:06 by chajung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../inc/libft.h"
+#include "../inc/libft.h"
 
-void	ft_putstr_fd(char *s, int fd)
+void	ft_striteri(char *s, void (*f)(unsigned int, char*))
 {
-	write(fd, s, ft_strlen(s));
+	unsigned int	i;
+
+	i = 0;
+	while (*s)
+		f(i++, s++);
 }
