+++ libft/src/ft_strjoin.c
@@ -1,22 +1,31 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_lstiter.c                                       :+:      :+:    :+:   */
+/*   ft_strjoin.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: chajung <chajung@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/11/12 16:42:15 by chajung           #+#    #+#             */
-/*   Updated: 2022/11/13 14:27:12 by chajung          ###   ########.fr       */
+/*   Created: 2022/11/10 12:56:27 by chajung           #+#    #+#             */
+/*   Updated: 2022/11/13 14:08:26 by chajung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../inc/libft.h"
+#include "../inc/libft.h"
 
-void	ft_lstiter(t_list *lst, void (*f)(void *))
+char	*ft_strjoin(char const *s1, char const *s2)
 {
-	while (lst)
-	{
-		f(lst->content);
-		lst = lst->next;
-	}
+	int		i;
+	char	*addr;
+
+	addr = (char *)malloc(sizeof(char) \
+		* (ft_strlen(s1) + ft_strlen(s2) + 1));
+	if (addr == NULL)
+		return (NULL);
+	i = 0;
+	while (*s1)
+		addr[i++] = *s1++;
+	while (*s2)
+		addr[i++] = *s2++;
+	addr[i] = '\0';
+	return (addr);
 }
