+++ libft/src/ft_strmapi.c
@@ -1,19 +1,31 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_lstadd_front.c                                  :+:      :+:    :+:   */
+/*   ft_strmapi.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: chajung <chajung@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/11/12 16:39:37 by chajung           #+#    #+#             */
-/*   Updated: 2022/11/13 14:27:09 by chajung          ###   ########.fr       */
+/*   Created: 2022/11/10 14:39:26 by chajung           #+#    #+#             */
+/*   Updated: 2022/11/13 14:08:44 by chajung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../inc/libft.h"
+#include "../inc/libft.h"
 
-void	ft_lstadd_front(t_list **lst, t_list *new)
+char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	new->next = *lst;
-	*lst = new;
+	unsigned int	i;
+	char			*addr;
+
+	addr = (char *)malloc(sizeof(char) * (ft_strlen(s) + 1));
+	if (addr == NULL)
+		return (NULL);
+	i = 0;
+	while (s[i])
+	{
+		addr[i] = f(i, s[i]);
+		i++;
+	}
+	addr[i] = '\0';
+	return (addr);
 }
