+++ libft/src/ft_strncmp.c
