+++ libft/src/ft_strrchr.c
@@ -1,28 +1,29 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_calloc.c                                        :+:      :+:    :+:   */
+/*   ft_strrchr.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: chajung <chajung@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/11/09 20:10:09 by chajung           #+#    #+#             */
-/*   Updated: 2022/11/13 16:51:16 by chajung          ###   ########.fr       */
+/*   Created: 2022/11/09 18:42:01 by chajung           #+#    #+#             */
+/*   Updated: 2022/11/13 14:09:08 by chajung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../inc/libft.h"
+#include "../inc/libft.h"
 
-void	*ft_calloc(size_t count, size_t size)
+char	*ft_strrchr(const char *s, int c)
 {
-	size_t	bytes;
-	void	*addr;
+	char	*temp;
 
-	bytes = count * size;
-	if (count != 0 && bytes / count != size)
-		return (NULL);
-	addr = malloc(bytes);
-	if (addr == NULL)
-		return (NULL);
-	ft_bzero(addr, bytes);
-	return (addr);
+	temp = NULL;
+	if ((char)c == '\0')
+		return ((char *)s + ft_strlen(s));
+	while (*s)
+	{
+		if (*s == (char)c)
+			temp = (char *)s;
+		s++;
+	}
+	return (temp);
 }
